extract to ConvenienceMethodTemplate
