+++ javagen/src/main/java/com/azure/autorest/template/ServiceAsyncClientTemplate.java
@@ -3,38 +3,24 @@
 
 package com.azure.autorest.template;
 
-import com.azure.autorest.extension.base.model.codemodel.RequestParameterLocation;
 import com.azure.autorest.extension.base.plugin.JavaSettings;
 import com.azure.autorest.model.clientmodel.AsyncSyncClient;
 import com.azure.autorest.model.clientmodel.ClassType;
 import com.azure.autorest.model.clientmodel.ClientBuilder;
-import com.azure.autorest.model.clientmodel.ClientMethod;
-import com.azure.autorest.model.clientmodel.ClientMethodParameter;
 import com.azure.autorest.model.clientmodel.ConvenienceMethod;
-import com.azure.autorest.model.clientmodel.EnumType;
-import com.azure.autorest.model.clientmodel.GenericType;
-import com.azure.autorest.model.clientmodel.IType;
 import com.azure.autorest.model.clientmodel.MethodGroupClient;
-import com.azure.autorest.model.clientmodel.ParameterSynthesizedOrigin;
-import com.azure.autorest.model.clientmodel.ProxyMethodParameter;
 import com.azure.autorest.model.clientmodel.ServiceClient;
 import com.azure.autorest.model.javamodel.JavaClass;
 import com.azure.autorest.model.javamodel.JavaContext;
 import com.azure.autorest.model.javamodel.JavaFile;
 import com.azure.autorest.model.javamodel.JavaVisibility;
 import com.azure.autorest.util.ClientModelUtil;
-import com.azure.autorest.util.CodeNamer;
 import com.azure.autorest.util.ModelNamer;
 import com.azure.core.client.traits.EndpointTrait;
 
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 /**
  * Template to create an asynchronous client.
@@ -135,7 +121,7 @@ public class ServiceAsyncClientTemplate implements IJavaTemplate<AsyncSyncClient
               Templates.getWrapperClientMethodTemplate().write(clientMethod, classBlock);
             });
 
-        asyncClient.getConvenienceMethods().forEach(m -> writeConvenienceMethods(classBlock, m));
+        asyncClient.getConvenienceMethods().forEach(m -> writeConvenienceMethods(m, classBlock));
       }
 
       ServiceAsyncClientTemplate.addEndpointMethod(classBlock, asyncClient.getClientBuilder(), "this.serviceClient");
@@ -189,159 +175,7 @@ public class ServiceAsyncClientTemplate implements IJavaTemplate<AsyncSyncClient
     ClassType.RequestOptions.addImportsTo(imports, false);
   }
 
-  private static void writeConvenienceMethods(JavaClass classBlock, ConvenienceMethod convenienceMethodObj) {
-    ClientMethod clientMethod = convenienceMethodObj.getClientMethod();
-    convenienceMethodObj.getConvenienceMethods().stream()
-        .filter(m -> m.getType().name().contains("Async"))
-        .forEach(convenienceMethod -> {
-          classBlock.javadocComment(comment -> {
-            ClientMethodTemplate.generateJavadoc(convenienceMethod, comment, convenienceMethod.getProxyMethod(), true);
-          });
-
-          classBlock.publicMethod(convenienceMethod.getDeclaration(), methodBlock -> {
-            // RequestOptions
-            methodBlock.line("RequestOptions requestOptions = new RequestOptions();");
-
-            Map<MethodParameter, MethodParameter> parametersMap =
-                findParametersForConvenienceMethod(convenienceMethod, clientMethod);
-            Map<String, String> parameterExpressionsMap = new HashMap<>();
-            for (Map.Entry<MethodParameter, MethodParameter> entry : parametersMap.entrySet()) {
-              MethodParameter parameter = entry.getKey();
-              MethodParameter clientParameter = entry.getValue();
-
-              if (parameter.getProxyMethodParameter().getOrigin() == ParameterSynthesizedOrigin.CONTEXT) {
-                // Context
-                methodBlock.line(String.format("requestOptions.setContext(%s);", parameter.getName()));
-              } else if (clientParameter != null) {
-                String expression = parameter.getName();
-                if (parameter.getProxyMethodParameter().getRequestParameterLocation() == RequestParameterLocation.BODY) {
-                  expression = expressionConvertToBinaryData(parameter.getName(), parameter.getClientMethodParameter().getClientType());
-                }
-                parameterExpressionsMap.put(clientParameter.getName(), expression);
-              } else {
-                switch (parameter.getProxyMethodParameter().getRequestParameterLocation()) {
-                  case HEADER:
-                    methodBlock.line(
-                            String.format("requestOptions.setHeader(%1$s, %2$s);",
-                                    ClassType.String.defaultValueExpression(parameter.getSerializedName()),
-                                    expressionConvertToString(parameter.getName(), parameter.getClientMethodParameter().getClientType())));
-                    break;
-
-                  case QUERY:
-                    methodBlock.line(
-                            String.format("requestOptions.addQueryParam(%1$s, %2$s);",
-                                    ClassType.String.defaultValueExpression(parameter.getSerializedName()),
-                                    expressionConvertToString(parameter.getName(), parameter.getClientMethodParameter().getClientType())));
-                    break;
-
-                  case BODY:
-                    methodBlock.line(
-                            String.format("requestOptions.setBody(%s);",
-                                    expressionConvertToBinaryData(parameter.getName(), parameter.getClientMethodParameter().getClientType())));
-                    break;
-                }
-              }
-            }
-
-            IType baseReturnType = ((GenericType) convenienceMethod.getReturnValue().getType()).getTypeArguments()[0];
-            String methodName = clientMethod.getName().endsWith("Async")
-                ? clientMethod.getName().substring(0, clientMethod.getName().length() - "Async".length())
-                : clientMethod.getName();
-            String invocationExpression = clientMethod.getMethodInputParameters().stream()
-                .map(p -> {
-                  String expression = parameterExpressionsMap.get(p.getName());
-                  if (expression == null) {
-                    expression = p.getName();
-                  }
-                  return expression;
-                })
-                .collect(Collectors.joining(", "));
-            String returnTypeConversionExpression = ClientModelUtil.isClientModel(baseReturnType)
-                ? String.format(".map(r -> r.toObject(%s.class))", baseReturnType)
-                : "";
-
-            methodBlock.methodReturn(
-                String.format("%1$s(%2$s).map(Response::getValue)%3$s",
-                    methodName,
-                    invocationExpression,
-                    returnTypeConversionExpression));
-          });
-        });
-  }
-
-  private static String expressionConvertToBinaryData(String name, IType type) {
-    if (type == ClassType.BinaryData) {
-      return name;
-    } else {
-      return String.format("BinaryData.fromObject(%s)", name);
-    }
-  }
-
-  private static String expressionConvertToString(String name, IType type) {
-    if (type == ClassType.String) {
-      return name;
-    } else if (type instanceof EnumType) {
-      return String.format("%s.toString()", name);
-    } else {
-      return String.format("String.valueOf(%s)", name);
-    }
-  }
-
-  private static Map<MethodParameter, MethodParameter> findParametersForConvenienceMethod(
-          ClientMethod convenienceMethod, ClientMethod clientMethod) {
-    Map<MethodParameter, MethodParameter> parameterMap = new LinkedHashMap<>();
-    List<MethodParameter> convenienceParameters = getParameters(convenienceMethod, true);
-    Map<String, MethodParameter> clientParameters = getParameters(clientMethod, false).stream()
-        .collect(Collectors.toMap(MethodParameter::getSerializedName, Function.identity()));
-    for (MethodParameter convenienceParameter : convenienceParameters) {
-      String name = convenienceParameter.getSerializedName();
-      parameterMap.put(convenienceParameter, clientParameters.get(name));
-    }
-    return parameterMap;
-  }
-
-  private static List<MethodParameter> getParameters(ClientMethod clientMethod, boolean useAllParameters) {
-    List<ProxyMethodParameter> proxyMethodParameters = useAllParameters ? clientMethod.getProxyMethod().getAllParameters() : clientMethod.getProxyMethod().getParameters();
-    Map<String, ProxyMethodParameter> proxyMethodParameterByClientParameterName = proxyMethodParameters.stream()
-            .collect(Collectors.toMap(p -> CodeNamer.getEscapedReservedClientMethodParameterName(p.getName()), Function.identity()));
-    return clientMethod.getMethodParameters().stream()
-        .filter(p -> !p.getIsConstant() && !p.getFromClient())
-        .map(p -> new MethodParameter(proxyMethodParameterByClientParameterName.get(p.getName()), p))
-        .collect(Collectors.toList());
-  }
-
-  private static class MethodParameter {
-
-    private final ProxyMethodParameter proxyMethodParameter;
-    private final ClientMethodParameter clientMethodParameter;
-
-    public MethodParameter(ProxyMethodParameter proxyMethodParameter, ClientMethodParameter clientMethodParameter) {
-      this.proxyMethodParameter = proxyMethodParameter;
-      this.clientMethodParameter = clientMethodParameter;
-    }
-
-    public ProxyMethodParameter getProxyMethodParameter() {
-      return proxyMethodParameter;
-    }
-
-    public ClientMethodParameter getClientMethodParameter() {
-      return clientMethodParameter;
-    }
-
-    public String getName() {
-      return this.getClientMethodParameter().getName();
-    }
-
-    public String getSerializedName() {
-      if (this.getProxyMethodParameter() == null) {
-        return null;
-      } else {
-        String name = this.getProxyMethodParameter().getRequestParameterName();
-        if (name == null && this.getProxyMethodParameter().getRequestParameterLocation() == RequestParameterLocation.BODY) {
-          name = "__internal_request_BODY";
-        }
-        return name;
-      }
-    }
+  private static void writeConvenienceMethods(ConvenienceMethod convenienceMethodObj, JavaClass classBlock) {
+    Templates.getConvenienceMethodTemplate().write(convenienceMethodObj, classBlock);
   }
 }
